Hay un error sigo trabajando.

+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/data/local/category/CategoryEntity.kt
@@ -13,7 +13,7 @@ data class CategoryEntity(
 
 fun CategoryEntity.asModel() =
     CategoryModel(
-        id = categoryId.toUInt(),
+        id = categoryId.toULong(),
         name = name,
     )
