+++ app/src/main/java/ar/edu/unlam/mobile/scaffolding/domain/shoppinglist/ShoppingListRepository.kt
@@ -15,7 +15,7 @@ interface ShoppingListRepository {
     /**
      * Obtiene una lista de compras por su ID.
      */
-    fun getShoppingListStream(id: Int): Flow<ShoppingListModel>
+    fun getShoppingListStream(id: Long): Flow<ShoppingListModel>
 
     /**
      * Inserta una nueva lista de compras.
